Move EnemyStrategy from Koga to Sato
