+++ include/Sato/EnemyStrategy.hpp
